Move fake cgroup manager into separate file
