+++ src/ray/common/cgroup2/cgroup_test_utils.h
@@ -15,13 +15,11 @@
 
 #include <sys/types.h>
 
-#include <cstdint>
 #include <fstream>
 #include <memory>
 #include <string>
 #include <utility>
 
-#include "ray/common/cgroup2/cgroup_manager_interface.h"
 #include "ray/common/status.h"
 #include "ray/common/status_or.h"
 
@@ -134,68 +132,3 @@ std::ostream &operator<<(std::ostream &os,
   }
   return os << *ptr;
 }
-
-namespace ray {
-
-/**
- * @brief A test-only fake CgroupManager that stores all cgroup state in a temporary
- * directory on tmpfs. Mocks memory.events and memory.stat
- */
-class FakeCgroupManager : public CgroupManagerInterface {
- public:
-  explicit FakeCgroupManager(int64_t user_memory_max_bytes,
-                             int64_t user_memory_high_bytes)
-      : user_memory_max_bytes_(user_memory_max_bytes),
-        user_memory_high_bytes_(user_memory_high_bytes) {
-    StatusOr<std::unique_ptr<TempDirectory>> temp_dir_or = TempDirectory::Create();
-    RAY_CHECK(temp_dir_or.ok()) << temp_dir_or.status().ToString();
-    temp_dir_ = std::move(temp_dir_or.value());
-
-    memory_events_file_ =
-        std::make_unique<TempFile>(temp_dir_->GetPath() + "/memory.events");
-    memory_events_file_->AppendLine("low 0\n");
-    memory_events_file_->AppendLine("high 0\n");
-    memory_events_file_->AppendLine("max 0\n");
-    memory_events_file_->AppendLine("oom 0\n");
-    memory_events_file_->AppendLine("oom_kill 0\n");
-
-    memory_stat_file_ = std::make_unique<TempFile>(temp_dir_->GetPath() + "/memory.stat");
-    memory_stat_file_->AppendLine("anon 0\n");
-    memory_stat_file_->AppendLine("shmem 0\n");
-  }
-
-  Status AddProcessToWorkersCgroup(const std::string &) override { return Status::OK(); }
-  Status AddProcessToSystemCgroup(const std::string &) override { return Status::OK(); }
-
-  std::string GetUserCgroupPath() const override { return temp_dir_->GetPath(); }
-  std::string GetSystemCgroupPath() const override { return temp_dir_->GetPath(); }
-
-  StatusOr<std::string> GetSystemCgroupConstraintValue(
-      const std::string &) const override {
-    return Status::IOError("not implemented");
-  }
-
-  StatusOr<std::string> GetUserCgroupConstraintValue(
-      const std::string &constraint_name) const override {
-    if (constraint_name == "memory.max") {
-      return std::to_string(user_memory_max_bytes_);
-    }
-    if (constraint_name == "memory.high") {
-      return std::to_string(user_memory_high_bytes_);
-    }
-    return Status::IOError("constraint not found: " + constraint_name);
-  }
-
-  const std::string &GetPath() const { return temp_dir_->GetPath(); }
-
- private:
-  // temp_dir_ must be declared before the TempFile members so that it is
-  // destroyed last, after the files are unlinked by their own destructors.
-  std::unique_ptr<TempDirectory> temp_dir_;
-  std::unique_ptr<TempFile> memory_events_file_;
-  std::unique_ptr<TempFile> memory_stat_file_;
-  int64_t user_memory_max_bytes_;
-  int64_t user_memory_high_bytes_;
-};
-
-}  // namespace ray
