+++ src/ray/common/tests/BUILD.bazel
@@ -196,7 +196,7 @@ ray_cc_test(
         "//src/ray/common:memory_monitor_factory",
         "//src/ray/common:memory_monitor_interface",
         "//src/ray/common:threshold_memory_monitor",
-        "//src/ray/common/cgroup2:cgroup_test_utils",
+        "//src/ray/common/cgroup2:fake_cgroup_manager",
     ],
 )
